+++ hexstrike_mcp.py
@@ -266,553 +266,18 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
     # Register enhanced web application security tools
     from mcp_tools.enhanced_web_app_security import register_enhanced_web_app_security_tools
     register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger)
-    
-    # ============================================================================
-    # AI-POWERED PAYLOAD GENERATION (v5.0 ENHANCEMENT)
-    # ============================================================================
 
-    @mcp.tool()
-    def ai_generate_payload(attack_type: str, complexity: str = "basic", technology: str = "", url: str = "") -> Dict[str, Any]:
-        """
-        Generate AI-powered contextual payloads for security testing.
+    # Register AI-powered payload generation and testing tools
+    from mcp_tools.ai_payload_generation import register_ai_payload_generation_tools
+    register_ai_payload_generation_tools(mcp, hexstrike_client, logger)
 
-        Args:
-            attack_type: Type of attack (xss, sqli, lfi, cmd_injection, ssti, xxe)
-            complexity: Complexity level (basic, advanced, bypass)
-            technology: Target technology (php, asp, jsp, python, nodejs)
-            url: Target URL for context
+    # Register API testing tools for comprehensive API security assessment
+    from mcp_tools.api_testing import register_api_testing_tools
+    register_api_testing_tools(mcp, hexstrike_client, logger)
 
-        Returns:
-            Contextual payloads with risk assessment and test cases
-        """
-        data = {
-            "attack_type": attack_type,
-            "complexity": complexity,
-            "technology": technology,
-            "url": url
-        }
-        logger.info(f"🤖 Generating AI payloads for {attack_type} attack")
-        result = hexstrike_client.safe_post("api/ai/generate_payload", data)
-
-        if result.get("success"):
-            payload_data = result.get("ai_payload_generation", {})
-            count = payload_data.get("payload_count", 0)
-            logger.info(f"✅ Generated {count} contextual {attack_type} payloads")
-
-            # Log some example payloads for user awareness
-            payloads = payload_data.get("payloads", [])
-            if payloads:
-                logger.info("🎯 Sample payloads generated:")
-                for i, payload_info in enumerate(payloads[:3]):  # Show first 3
-                    risk = payload_info.get("risk_level", "UNKNOWN")
-                    context = payload_info.get("context", "basic")
-                    logger.info(f"   ├─ [{risk}] {context}: {payload_info['payload'][:50]}...")
-        else:
-            logger.error("❌ AI payload generation failed")
-
-        return result
-
-    @mcp.tool()
-    def ai_test_payload(payload: str, target_url: str, method: str = "GET") -> Dict[str, Any]:
-        """
-        Test generated payload against target with AI analysis.
-
-        Args:
-            payload: The payload to test
-            target_url: Target URL to test against
-            method: HTTP method (GET, POST)
-
-        Returns:
-            Test results with AI analysis and vulnerability assessment
-        """
-        data = {
-            "payload": payload,
-            "target_url": target_url,
-            "method": method
-        }
-        logger.info(f"🧪 Testing AI payload against {target_url}")
-        result = hexstrike_client.safe_post("api/ai/test_payload", data)
-
-        if result.get("success"):
-            analysis = result.get("ai_analysis", {})
-            potential_vuln = analysis.get("potential_vulnerability", False)
-            logger.info(f"🔍 Payload test completed | Vulnerability detected: {potential_vuln}")
-
-            if potential_vuln:
-                logger.warning("⚠️  Potential vulnerability found! Review the response carefully.")
-            else:
-                logger.info("✅ No obvious vulnerability indicators detected")
-        else:
-            logger.error("❌ Payload testing failed")
-
-        return result
-
-    @mcp.tool()
-    def ai_generate_attack_suite(target_url: str, attack_types: str = "xss,sqli,lfi") -> Dict[str, Any]:
-        """
-        Generate comprehensive attack suite with multiple payload types.
-
-        Args:
-            target_url: Target URL for testing
-            attack_types: Comma-separated list of attack types
-
-        Returns:
-            Comprehensive attack suite with multiple payload types
-        """
-        attack_list = [attack.strip() for attack in attack_types.split(",")]
-        results = {
-            "target_url": target_url,
-            "attack_types": attack_list,
-            "payload_suites": {},
-            "summary": {
-                "total_payloads": 0,
-                "high_risk_payloads": 0,
-                "test_cases": 0
-            }
-        }
-
-        logger.info(f"🚀 Generating comprehensive attack suite for {target_url}")
-        logger.info(f"🎯 Attack types: {', '.join(attack_list)}")
-
-        for attack_type in attack_list:
-            logger.info(f"🤖 Generating {attack_type} payloads...")
-
-            # Generate payloads for this attack type
-            payload_result = ai_generate_payload(attack_type, "advanced", "", target_url)
-
-            if payload_result.get("success"):
-                payload_data = payload_result.get("ai_payload_generation", {})
-                results["payload_suites"][attack_type] = payload_data
-
-                # Update summary
-                results["summary"]["total_payloads"] += payload_data.get("payload_count", 0)
-                results["summary"]["test_cases"] += len(payload_data.get("test_cases", []))
-
-                # Count high-risk payloads
-                for payload_info in payload_data.get("payloads", []):
-                    if payload_info.get("risk_level") == "HIGH":
-                        results["summary"]["high_risk_payloads"] += 1
-
-        logger.info(f"✅ Attack suite generated:")
-        logger.info(f"   ├─ Total payloads: {results['summary']['total_payloads']}")
-        logger.info(f"   ├─ High-risk payloads: {results['summary']['high_risk_payloads']}")
-        logger.info(f"   └─ Test cases: {results['summary']['test_cases']}")
-
-        return {
-            "success": True,
-            "attack_suite": results,
-            "timestamp": time.time()
-        }
-
-    # ============================================================================
-    # ADVANCED API TESTING TOOLS (v5.0 ENHANCEMENT)
-    # ============================================================================
-
-    @mcp.tool()
-    def api_fuzzer(base_url: str, endpoints: str = "", methods: str = "GET,POST,PUT,DELETE", wordlist: str = "/usr/share/wordlists/api/api-endpoints.txt") -> Dict[str, Any]:
-        """
-        Advanced API endpoint fuzzing with intelligent parameter discovery.
-
-        Args:
-            base_url: Base URL of the API
-            endpoints: Comma-separated list of specific endpoints to test
-            methods: HTTP methods to test (comma-separated)
-            wordlist: Wordlist for endpoint discovery
-
-        Returns:
-            API fuzzing results with endpoint discovery and vulnerability assessment
-        """
-        data = {
-            "base_url": base_url,
-            "endpoints": [e.strip() for e in endpoints.split(",") if e.strip()] if endpoints else [],
-            "methods": [m.strip() for m in methods.split(",")],
-            "wordlist": wordlist
-        }
-
-        logger.info(f"🔍 Starting API fuzzing: {base_url}")
-        result = hexstrike_client.safe_post("api/tools/api_fuzzer", data)
-
-        if result.get("success"):
-            fuzzing_type = result.get("fuzzing_type", "unknown")
-            if fuzzing_type == "endpoint_testing":
-                endpoint_count = len(result.get("results", []))
-                logger.info(f"✅ API endpoint testing completed: {endpoint_count} endpoints tested")
-            else:
-                logger.info(f"✅ API endpoint discovery completed")
-        else:
-            logger.error("❌ API fuzzing failed")
-
-        return result
-
-    @mcp.tool()
-    def graphql_scanner(endpoint: str, introspection: bool = True, query_depth: int = 10, test_mutations: bool = True) -> Dict[str, Any]:
-        """
-        Advanced GraphQL security scanning and introspection.
-
-        Args:
-            endpoint: GraphQL endpoint URL
-            introspection: Test introspection queries
-            query_depth: Maximum query depth to test
-            test_mutations: Test mutation operations
-
-        Returns:
-            GraphQL security scan results with vulnerability assessment
-        """
-        data = {
-            "endpoint": endpoint,
-            "introspection": introspection,
-            "query_depth": query_depth,
-            "test_mutations": test_mutations
-        }
-
-        logger.info(f"🔍 Starting GraphQL security scan: {endpoint}")
-        result = hexstrike_client.safe_post("api/tools/graphql_scanner", data)
-
-        if result.get("success"):
-            scan_results = result.get("graphql_scan_results", {})
-            vuln_count = len(scan_results.get("vulnerabilities", []))
-            tests_count = len(scan_results.get("tests_performed", []))
-
-            logger.info(f"✅ GraphQL scan completed: {tests_count} tests, {vuln_count} vulnerabilities")
-
-            if vuln_count > 0:
-                logger.warning(f"⚠️  Found {vuln_count} GraphQL vulnerabilities!")
-                for vuln in scan_results.get("vulnerabilities", [])[:3]:  # Show first 3
-                    severity = vuln.get("severity", "UNKNOWN")
-                    vuln_type = vuln.get("type", "unknown")
-                    logger.warning(f"   ├─ [{severity}] {vuln_type}")
-        else:
-            logger.error("❌ GraphQL scanning failed")
-
-        return result
-
-    @mcp.tool()
-    def jwt_analyzer(jwt_token: str, target_url: str = "") -> Dict[str, Any]:
-        """
-        Advanced JWT token analysis and vulnerability testing.
-
-        Args:
-            jwt_token: JWT token to analyze
-            target_url: Optional target URL for testing token manipulation
-
-        Returns:
-            JWT analysis results with vulnerability assessment and attack vectors
-        """
-        data = {
-            "jwt_token": jwt_token,
-            "target_url": target_url
-        }
-
-        logger.info(f"🔍 Starting JWT security analysis")
-        result = hexstrike_client.safe_post("api/tools/jwt_analyzer", data)
-
-        if result.get("success"):
-            analysis = result.get("jwt_analysis_results", {})
-            vuln_count = len(analysis.get("vulnerabilities", []))
-            algorithm = analysis.get("token_info", {}).get("algorithm", "unknown")
-
-            logger.info(f"✅ JWT analysis completed: {vuln_count} vulnerabilities found")
-            logger.info(f"🔐 Token algorithm: {algorithm}")
-
-            if vuln_count > 0:
-                logger.warning(f"⚠️  Found {vuln_count} JWT vulnerabilities!")
-                for vuln in analysis.get("vulnerabilities", [])[:3]:  # Show first 3
-                    severity = vuln.get("severity", "UNKNOWN")
-                    vuln_type = vuln.get("type", "unknown")
-                    logger.warning(f"   ├─ [{severity}] {vuln_type}")
-        else:
-            logger.error("❌ JWT analysis failed")
-
-        return result
-
-    @mcp.tool()
-    def api_schema_analyzer(schema_url: str, schema_type: str = "openapi") -> Dict[str, Any]:
-        """
-        Analyze API schemas and identify potential security issues.
-
-        Args:
-            schema_url: URL to the API schema (OpenAPI/Swagger/GraphQL)
-            schema_type: Type of schema (openapi, swagger, graphql)
-
-        Returns:
-            Schema analysis results with security issues and recommendations
-        """
-        data = {
-            "schema_url": schema_url,
-            "schema_type": schema_type
-        }
-
-        logger.info(f"🔍 Starting API schema analysis: {schema_url}")
-        result = hexstrike_client.safe_post("api/tools/api_schema_analyzer", data)
-
-        if result.get("success"):
-            analysis = result.get("schema_analysis_results", {})
-            endpoint_count = len(analysis.get("endpoints_found", []))
-            issue_count = len(analysis.get("security_issues", []))
-
-            logger.info(f"✅ Schema analysis completed: {endpoint_count} endpoints, {issue_count} issues")
-
-            if issue_count > 0:
-                logger.warning(f"⚠️  Found {issue_count} security issues in schema!")
-                for issue in analysis.get("security_issues", [])[:3]:  # Show first 3
-                    severity = issue.get("severity", "UNKNOWN")
-                    issue_type = issue.get("issue", "unknown")
-                    logger.warning(f"   ├─ [{severity}] {issue_type}")
-
-            if endpoint_count > 0:
-                logger.info(f"📊 Discovered endpoints:")
-                for endpoint in analysis.get("endpoints_found", [])[:5]:  # Show first 5
-                    method = endpoint.get("method", "GET")
-                    path = endpoint.get("path", "/")
-                    logger.info(f"   ├─ {method} {path}")
-        else:
-            logger.error("❌ Schema analysis failed")
-
-        return result
-
-    @mcp.tool()
-    def comprehensive_api_audit(base_url: str, schema_url: str = "", jwt_token: str = "", graphql_endpoint: str = "") -> Dict[str, Any]:
-        """
-        Comprehensive API security audit combining multiple testing techniques.
-
-        Args:
-            base_url: Base URL of the API
-            schema_url: Optional API schema URL
-            jwt_token: Optional JWT token for analysis
-            graphql_endpoint: Optional GraphQL endpoint
-
-        Returns:
-            Comprehensive audit results with all API security tests
-        """
-        audit_results = {
-            "base_url": base_url,
-            "audit_timestamp": time.time(),
-            "tests_performed": [],
-            "total_vulnerabilities": 0,
-            "summary": {},
-            "recommendations": []
-        }
-
-        logger.info(f"🚀 Starting comprehensive API security audit: {base_url}")
-
-        # 1. API Endpoint Fuzzing
-        logger.info("🔍 Phase 1: API endpoint discovery and fuzzing")
-        fuzz_result = api_fuzzer(base_url)
-        if fuzz_result.get("success"):
-            audit_results["tests_performed"].append("api_fuzzing")
-            audit_results["api_fuzzing"] = fuzz_result
-
-        # 2. Schema Analysis (if provided)
-        if schema_url:
-            logger.info("🔍 Phase 2: API schema analysis")
-            schema_result = api_schema_analyzer(schema_url)
-            if schema_result.get("success"):
-                audit_results["tests_performed"].append("schema_analysis")
-                audit_results["schema_analysis"] = schema_result
-
-                schema_data = schema_result.get("schema_analysis_results", {})
-                audit_results["total_vulnerabilities"] += len(schema_data.get("security_issues", []))
-
-        # 3. JWT Analysis (if provided)
-        if jwt_token:
-            logger.info("🔍 Phase 3: JWT token analysis")
-            jwt_result = jwt_analyzer(jwt_token, base_url)
-            if jwt_result.get("success"):
-                audit_results["tests_performed"].append("jwt_analysis")
-                audit_results["jwt_analysis"] = jwt_result
-
-                jwt_data = jwt_result.get("jwt_analysis_results", {})
-                audit_results["total_vulnerabilities"] += len(jwt_data.get("vulnerabilities", []))
-
-        # 4. GraphQL Testing (if provided)
-        if graphql_endpoint:
-            logger.info("🔍 Phase 4: GraphQL security scanning")
-            graphql_result = graphql_scanner(graphql_endpoint)
-            if graphql_result.get("success"):
-                audit_results["tests_performed"].append("graphql_scanning")
-                audit_results["graphql_scanning"] = graphql_result
-
-                graphql_data = graphql_result.get("graphql_scan_results", {})
-                audit_results["total_vulnerabilities"] += len(graphql_data.get("vulnerabilities", []))
-
-        # Generate comprehensive recommendations
-        audit_results["recommendations"] = [
-            "Implement proper authentication and authorization",
-            "Use HTTPS for all API communications",
-            "Validate and sanitize all input parameters",
-            "Implement rate limiting and request throttling",
-            "Add comprehensive logging and monitoring",
-            "Regular security testing and code reviews",
-            "Keep API documentation updated and secure",
-            "Implement proper error handling"
-        ]
-
-        # Summary
-        audit_results["summary"] = {
-            "tests_performed": len(audit_results["tests_performed"]),
-            "total_vulnerabilities": audit_results["total_vulnerabilities"],
-            "audit_coverage": "comprehensive" if len(audit_results["tests_performed"]) >= 3 else "partial"
-        }
-
-        logger.info(f"✅ Comprehensive API audit completed:")
-        logger.info(f"   ├─ Tests performed: {audit_results['summary']['tests_performed']}")
-        logger.info(f"   ├─ Total vulnerabilities: {audit_results['summary']['total_vulnerabilities']}")
-        logger.info(f"   └─ Coverage: {audit_results['summary']['audit_coverage']}")
-
-        return {
-            "success": True,
-            "comprehensive_audit": audit_results
-        }
-
-    # ============================================================================
-    # ADVANCED CTF TOOLS (v5.0 ENHANCEMENT)
-    # ============================================================================
-
-    @mcp.tool()
-    def volatility3_analyze(memory_file: str, plugin: str, output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Volatility3 for advanced memory forensics with enhanced logging.
-
-        Args:
-            memory_file: Path to memory dump file
-            plugin: Volatility3 plugin to execute
-            output_file: Output file path
-            additional_args: Additional Volatility3 arguments
-
-        Returns:
-            Advanced memory forensics results
-        """
-        data = {
-            "memory_file": memory_file,
-            "plugin": plugin,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🧠 Starting Volatility3 analysis: {plugin}")
-        result = hexstrike_client.safe_post("api/tools/volatility3", data)
-        if result.get("success"):
-            logger.info(f"✅ Volatility3 analysis completed")
-        else:
-            logger.error(f"❌ Volatility3 analysis failed")
-        return result
-
-    @mcp.tool()
-    def foremost_carving(input_file: str, output_dir: str = "/tmp/foremost_output", file_types: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Foremost for file carving with enhanced logging.
-
-        Args:
-            input_file: Input file or device to carve
-            output_dir: Output directory for carved files
-            file_types: File types to carve (jpg,gif,png,etc.)
-            additional_args: Additional Foremost arguments
-
-        Returns:
-            File carving results
-        """
-        data = {
-            "input_file": input_file,
-            "output_dir": output_dir,
-            "file_types": file_types,
-            "additional_args": additional_args
-        }
-        logger.info(f"📁 Starting Foremost file carving: {input_file}")
-        result = hexstrike_client.safe_post("api/tools/foremost", data)
-        if result.get("success"):
-            logger.info(f"✅ Foremost carving completed")
-        else:
-            logger.error(f"❌ Foremost carving failed")
-        return result
-
-    @mcp.tool()
-    def steghide_analysis(action: str, cover_file: str, embed_file: str = "", passphrase: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Steghide for steganography analysis with enhanced logging.
-
-        Args:
-            action: Action to perform (extract, embed, info)
-            cover_file: Cover file for steganography
-            embed_file: File to embed (for embed action)
-            passphrase: Passphrase for steganography
-            output_file: Output file path
-            additional_args: Additional Steghide arguments
-
-        Returns:
-            Steganography analysis results
-        """
-        data = {
-            "action": action,
-            "cover_file": cover_file,
-            "embed_file": embed_file,
-            "passphrase": passphrase,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🖼️ Starting Steghide {action}: {cover_file}")
-        result = hexstrike_client.safe_post("api/tools/steghide", data)
-        if result.get("success"):
-            logger.info(f"✅ Steghide {action} completed")
-        else:
-            logger.error(f"❌ Steghide {action} failed")
-        return result
-
-    @mcp.tool()
-    def exiftool_extract(file_path: str, output_format: str = "", tags: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute ExifTool for metadata extraction with enhanced logging.
-
-        Args:
-            file_path: Path to file for metadata extraction
-            output_format: Output format (json, xml, csv)
-            tags: Specific tags to extract
-            additional_args: Additional ExifTool arguments
-
-        Returns:
-            Metadata extraction results
-        """
-        data = {
-            "file_path": file_path,
-            "output_format": output_format,
-            "tags": tags,
-            "additional_args": additional_args
-        }
-        logger.info(f"📷 Starting ExifTool analysis: {file_path}")
-        result = hexstrike_client.safe_post("api/tools/exiftool", data)
-        if result.get("success"):
-            logger.info(f"✅ ExifTool analysis completed")
-        else:
-            logger.error(f"❌ ExifTool analysis failed")
-        return result
-
-    @mcp.tool()
-    def hashpump_attack(signature: str, data: str, key_length: str, append_data: str, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute HashPump for hash length extension attacks with enhanced logging.
-
-        Args:
-            signature: Original hash signature
-            data: Original data
-            key_length: Length of secret key
-            append_data: Data to append
-            additional_args: Additional HashPump arguments
-
-        Returns:
-            Hash length extension attack results
-        """
-        payload = {
-            "signature": signature,
-            "data": data,
-            "key_length": key_length,
-            "append_data": append_data,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔐 Starting HashPump attack")
-        result = hexstrike_client.safe_post("api/tools/hashpump", payload)
-        if result.get("success"):
-            logger.info(f"✅ HashPump attack completed")
-        else:
-            logger.error(f"❌ HashPump attack failed")
-        return result
+    # Register advanced CTF tools for competitive security challenges
+    from mcp_tools.advanced_ctf_tools import register_advanced_ctf_tools
+    register_advanced_ctf_tools(mcp, hexstrike_client, logger)
 
     # ============================================================================
     # BUG BOUNTY RECONNAISSANCE TOOLS (v5.0 ENHANCEMENT)
@@ -887,334 +352,6 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
             logger.info(f"✅ ParamSpider discovery completed")
         else:
             logger.error(f"❌ ParamSpider discovery failed")
-        return result
-
-    # ============================================================================
-    # ADVANCED WEB SECURITY TOOLS CONTINUED
-    # ============================================================================
-
-    @mcp.tool()
-    def burpsuite_scan(project_file: str = "", config_file: str = "", target: str = "", headless: bool = False, scan_type: str = "", scan_config: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Burp Suite with enhanced logging.
-
-        Args:
-            project_file: Burp project file path
-            config_file: Burp configuration file path
-            target: Target URL
-            headless: Run in headless mode
-            scan_type: Type of scan to perform
-            scan_config: Scan configuration
-            output_file: Output file path
-            additional_args: Additional Burp Suite arguments
-
-        Returns:
-            Burp Suite scan results
-        """
-        data = {
-            "project_file": project_file,
-            "config_file": config_file,
-            "target": target,
-            "headless": headless,
-            "scan_type": scan_type,
-            "scan_config": scan_config,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Burp Suite scan")
-        result = hexstrike_client.safe_post("api/tools/burpsuite", data)
-        if result.get("success"):
-            logger.info(f"✅ Burp Suite scan completed")
-        else:
-            logger.error(f"❌ Burp Suite scan failed")
-        return result
-
-    @mcp.tool()
-    def zap_scan(target: str = "", scan_type: str = "baseline", api_key: str = "", daemon: bool = False, port: str = "8090", host: str = "0.0.0.0", format_type: str = "xml", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute OWASP ZAP with enhanced logging.
-
-        Args:
-            target: Target URL
-            scan_type: Type of scan (baseline, full, api)
-            api_key: ZAP API key
-            daemon: Run in daemon mode
-            port: Port for ZAP daemon
-            host: Host for ZAP daemon
-            format_type: Output format (xml, json, html)
-            output_file: Output file path
-            additional_args: Additional ZAP arguments
-
-        Returns:
-            ZAP scan results
-        """
-        data = {
-            "target": target,
-            "scan_type": scan_type,
-            "api_key": api_key,
-            "daemon": daemon,
-            "port": port,
-            "host": host,
-            "format": format_type,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting ZAP scan: {target}")
-        result = hexstrike_client.safe_post("api/tools/zap", data)
-        if result.get("success"):
-            logger.info(f"✅ ZAP scan completed for {target}")
-        else:
-            logger.error(f"❌ ZAP scan failed for {target}")
-        return result
-
-    @mcp.tool()
-    def arjun_scan(url: str, method: str = "GET", data: str = "", headers: str = "", timeout: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Arjun for parameter discovery with enhanced logging.
-
-        Args:
-            url: Target URL
-            method: HTTP method (GET, POST, etc.)
-            data: POST data for testing
-            headers: Custom headers
-            timeout: Request timeout
-            output_file: Output file path
-            additional_args: Additional Arjun arguments
-
-        Returns:
-            Parameter discovery results
-        """
-        payload = {
-            "url": url,
-            "method": method,
-            "data": data,
-            "headers": headers,
-            "timeout": timeout,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Arjun parameter discovery: {url}")
-        result = hexstrike_client.safe_post("api/tools/arjun", payload)
-        if result.get("success"):
-            logger.info(f"✅ Arjun completed for {url}")
-        else:
-            logger.error(f"❌ Arjun failed for {url}")
-        return result
-
-    @mcp.tool()
-    def wafw00f_scan(target: str, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute wafw00f to identify and fingerprint WAF products with enhanced logging.
-
-        Args:
-            target: Target URL or IP
-            additional_args: Additional wafw00f arguments
-
-        Returns:
-            WAF detection results
-        """
-        data = {
-            "target": target,
-            "additional_args": additional_args
-        }
-        logger.info(f"🛡️ Starting Wafw00f WAF detection: {target}")
-        result = hexstrike_client.safe_post("api/tools/wafw00f", data)
-        if result.get("success"):
-            logger.info(f"✅ Wafw00f completed for {target}")
-        else:
-            logger.error(f"❌ Wafw00f failed for {target}")
-        return result
-
-    @mcp.tool()
-    def fierce_scan(domain: str, dns_server: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute fierce for DNS reconnaissance with enhanced logging.
-
-        Args:
-            domain: Target domain
-            dns_server: DNS server to use
-            additional_args: Additional fierce arguments
-
-        Returns:
-            DNS reconnaissance results
-        """
-        data = {
-            "domain": domain,
-            "dns_server": dns_server,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Fierce DNS recon: {domain}")
-        result = hexstrike_client.safe_post("api/tools/fierce", data)
-        if result.get("success"):
-            logger.info(f"✅ Fierce completed for {domain}")
-        else:
-            logger.error(f"❌ Fierce failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def dnsenum_scan(domain: str, dns_server: str = "", wordlist: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute dnsenum for DNS enumeration with enhanced logging.
-
-        Args:
-            domain: Target domain
-            dns_server: DNS server to use
-            wordlist: Wordlist for brute forcing
-            additional_args: Additional dnsenum arguments
-
-        Returns:
-            DNS enumeration results
-        """
-        data = {
-            "domain": domain,
-            "dns_server": dns_server,
-            "wordlist": wordlist,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting DNSenum: {domain}")
-        result = hexstrike_client.safe_post("api/tools/dnsenum", data)
-        if result.get("success"):
-            logger.info(f"✅ DNSenum completed for {domain}")
-        else:
-            logger.error(f"❌ DNSenum failed for {domain}")
-        return result
-
-    @mcp.tool()
-    def autorecon_scan(
-        target: str = "",
-        target_file: str = "",
-        ports: str = "",
-        output_dir: str = "",
-        max_scans: str = "",
-        max_port_scans: str = "",
-        heartbeat: str = "",
-        timeout: str = "",
-        target_timeout: str = "",
-        config_file: str = "",
-        global_file: str = "",
-        plugins_dir: str = "",
-        add_plugins_dir: str = "",
-        tags: str = "",
-        exclude_tags: str = "",
-        port_scans: str = "",
-        service_scans: str = "",
-        reports: str = "",
-        single_target: bool = False,
-        only_scans_dir: bool = False,
-        no_port_dirs: bool = False,
-        nmap: str = "",
-        nmap_append: str = "",
-        proxychains: bool = False,
-        disable_sanity_checks: bool = False,
-        disable_keyboard_control: bool = False,
-        force_services: str = "",
-        accessible: bool = False,
-        verbose: int = 0,
-        curl_path: str = "",
-        dirbuster_tool: str = "",
-        dirbuster_wordlist: str = "",
-        dirbuster_threads: str = "",
-        dirbuster_ext: str = "",
-        onesixtyone_community_strings: str = "",
-        global_username_wordlist: str = "",
-        global_password_wordlist: str = "",
-        global_domain: str = "",
-        additional_args: str = ""
-    ) -> Dict[str, Any]:
-        """
-        Execute AutoRecon for comprehensive target enumeration with full parameter support.
-
-        Args:
-            target: Single target to scan
-            target_file: File containing multiple targets
-            ports: Specific ports to scan
-            output_dir: Output directory
-            max_scans: Maximum number of concurrent scans
-            max_port_scans: Maximum number of concurrent port scans
-            heartbeat: Heartbeat interval
-            timeout: Global timeout
-            target_timeout: Per-target timeout
-            config_file: Configuration file path
-            global_file: Global configuration file
-            plugins_dir: Plugins directory
-            add_plugins_dir: Additional plugins directory
-            tags: Plugin tags to include
-            exclude_tags: Plugin tags to exclude
-            port_scans: Port scan plugins to run
-            service_scans: Service scan plugins to run
-            reports: Report plugins to run
-            single_target: Use single target directory structure
-            only_scans_dir: Only create scans directory
-            no_port_dirs: Don't create port directories
-            nmap: Custom nmap command
-            nmap_append: Arguments to append to nmap
-            proxychains: Use proxychains
-            disable_sanity_checks: Disable sanity checks
-            disable_keyboard_control: Disable keyboard control
-            force_services: Force service detection
-            accessible: Enable accessible output
-            verbose: Verbosity level (0-3)
-            curl_path: Custom curl path
-            dirbuster_tool: Directory busting tool
-            dirbuster_wordlist: Directory busting wordlist
-            dirbuster_threads: Directory busting threads
-            dirbuster_ext: Directory busting extensions
-            onesixtyone_community_strings: SNMP community strings
-            global_username_wordlist: Global username wordlist
-            global_password_wordlist: Global password wordlist
-            global_domain: Global domain
-            additional_args: Additional AutoRecon arguments
-
-        Returns:
-            Comprehensive enumeration results with full configurability
-        """
-        data = {
-            "target": target,
-            "target_file": target_file,
-            "ports": ports,
-            "output_dir": output_dir,
-            "max_scans": max_scans,
-            "max_port_scans": max_port_scans,
-            "heartbeat": heartbeat,
-            "timeout": timeout,
-            "target_timeout": target_timeout,
-            "config_file": config_file,
-            "global_file": global_file,
-            "plugins_dir": plugins_dir,
-            "add_plugins_dir": add_plugins_dir,
-            "tags": tags,
-            "exclude_tags": exclude_tags,
-            "port_scans": port_scans,
-            "service_scans": service_scans,
-            "reports": reports,
-            "single_target": single_target,
-            "only_scans_dir": only_scans_dir,
-            "no_port_dirs": no_port_dirs,
-            "nmap": nmap,
-            "nmap_append": nmap_append,
-            "proxychains": proxychains,
-            "disable_sanity_checks": disable_sanity_checks,
-            "disable_keyboard_control": disable_keyboard_control,
-            "force_services": force_services,
-            "accessible": accessible,
-            "verbose": verbose,
-            "curl_path": curl_path,
-            "dirbuster_tool": dirbuster_tool,
-            "dirbuster_wordlist": dirbuster_wordlist,
-            "dirbuster_threads": dirbuster_threads,
-            "dirbuster_ext": dirbuster_ext,
-            "onesixtyone_community_strings": onesixtyone_community_strings,
-            "global_username_wordlist": global_username_wordlist,
-            "global_password_wordlist": global_password_wordlist,
-            "global_domain": global_domain,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting AutoRecon comprehensive enumeration: {target}")
-        result = hexstrike_client.safe_post("api/tools/autorecon", data)
-        if result.get("success"):
-            logger.info(f"✅ AutoRecon comprehensive enumeration completed for {target}")
-        else:
-            logger.error(f"❌ AutoRecon failed for {target}")
         return result
 
     # ============================================================================
@@ -2337,8 +1474,6 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
             },
             "timestamp": datetime.now().isoformat()
         }
-
-    
 
     # ============================================================================
     # BUG BOUNTY HUNTING SPECIALIZED WORKFLOWS
