+++ mcp_tools/api_testing.py
@@ -1,0 +1,261 @@
+# mcp_tools/api_testing.py
+
+from typing import Dict, Any
+import time
+
+def register_api_testing_tools(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def api_fuzzer(base_url: str, endpoints: str = "", methods: str = "GET,POST,PUT,DELETE", wordlist: str = "/usr/share/wordlists/api/api-endpoints.txt") -> Dict[str, Any]:
+        """
+        Advanced API endpoint fuzzing with intelligent parameter discovery.
+
+        Args:
+            base_url: Base URL of the API
+            endpoints: Comma-separated list of specific endpoints to test
+            methods: HTTP methods to test (comma-separated)
+            wordlist: Wordlist for endpoint discovery
+
+        Returns:
+            API fuzzing results with endpoint discovery and vulnerability assessment
+        """
+        data = {
+            "base_url": base_url,
+            "endpoints": [e.strip() for e in endpoints.split(",") if e.strip()] if endpoints else [],
+            "methods": [m.strip() for m in methods.split(",")],
+            "wordlist": wordlist
+        }
+
+        logger.info(f"🔍 Starting API fuzzing: {base_url}")
+        result = hexstrike_client.safe_post("api/tools/api_fuzzer", data)
+
+        if result.get("success"):
+            fuzzing_type = result.get("fuzzing_type", "unknown")
+            if fuzzing_type == "endpoint_testing":
+                endpoint_count = len(result.get("results", []))
+                logger.info(f"✅ API endpoint testing completed: {endpoint_count} endpoints tested")
+            else:
+                logger.info(f"✅ API endpoint discovery completed")
+        else:
+            logger.error("❌ API fuzzing failed")
+
+        return result
+
+    @mcp.tool()
+    def graphql_scanner(endpoint: str, introspection: bool = True, query_depth: int = 10, test_mutations: bool = True) -> Dict[str, Any]:
+        """
+        Advanced GraphQL security scanning and introspection.
+
+        Args:
+            endpoint: GraphQL endpoint URL
+            introspection: Test introspection queries
+            query_depth: Maximum query depth to test
+            test_mutations: Test mutation operations
+
+        Returns:
+            GraphQL security scan results with vulnerability assessment
+        """
+        data = {
+            "endpoint": endpoint,
+            "introspection": introspection,
+            "query_depth": query_depth,
+            "test_mutations": test_mutations
+        }
+
+        logger.info(f"🔍 Starting GraphQL security scan: {endpoint}")
+        result = hexstrike_client.safe_post("api/tools/graphql_scanner", data)
+
+        if result.get("success"):
+            scan_results = result.get("graphql_scan_results", {})
+            vuln_count = len(scan_results.get("vulnerabilities", []))
+            tests_count = len(scan_results.get("tests_performed", []))
+
+            logger.info(f"✅ GraphQL scan completed: {tests_count} tests, {vuln_count} vulnerabilities")
+
+            if vuln_count > 0:
+                logger.warning(f"⚠️  Found {vuln_count} GraphQL vulnerabilities!")
+                for vuln in scan_results.get("vulnerabilities", [])[:3]:  # Show first 3
+                    severity = vuln.get("severity", "UNKNOWN")
+                    vuln_type = vuln.get("type", "unknown")
+                    logger.warning(f"   ├─ [{severity}] {vuln_type}")
+        else:
+            logger.error("❌ GraphQL scanning failed")
+
+        return result
+
+    @mcp.tool()
+    def jwt_analyzer(jwt_token: str, target_url: str = "") -> Dict[str, Any]:
+        """
+        Advanced JWT token analysis and vulnerability testing.
+
+        Args:
+            jwt_token: JWT token to analyze
+            target_url: Optional target URL for testing token manipulation
+
+        Returns:
+            JWT analysis results with vulnerability assessment and attack vectors
+        """
+        data = {
+            "jwt_token": jwt_token,
+            "target_url": target_url
+        }
+
+        logger.info(f"🔍 Starting JWT security analysis")
+        result = hexstrike_client.safe_post("api/tools/jwt_analyzer", data)
+
+        if result.get("success"):
+            analysis = result.get("jwt_analysis_results", {})
+            vuln_count = len(analysis.get("vulnerabilities", []))
+            algorithm = analysis.get("token_info", {}).get("algorithm", "unknown")
+
+            logger.info(f"✅ JWT analysis completed: {vuln_count} vulnerabilities found")
+            logger.info(f"🔐 Token algorithm: {algorithm}")
+
+            if vuln_count > 0:
+                logger.warning(f"⚠️  Found {vuln_count} JWT vulnerabilities!")
+                for vuln in analysis.get("vulnerabilities", [])[:3]:  # Show first 3
+                    severity = vuln.get("severity", "UNKNOWN")
+                    vuln_type = vuln.get("type", "unknown")
+                    logger.warning(f"   ├─ [{severity}] {vuln_type}")
+        else:
+            logger.error("❌ JWT analysis failed")
+
+        return result
+
+    @mcp.tool()
+    def api_schema_analyzer(schema_url: str, schema_type: str = "openapi") -> Dict[str, Any]:
+        """
+        Analyze API schemas and identify potential security issues.
+
+        Args:
+            schema_url: URL to the API schema (OpenAPI/Swagger/GraphQL)
+            schema_type: Type of schema (openapi, swagger, graphql)
+
+        Returns:
+            Schema analysis results with security issues and recommendations
+        """
+        data = {
+            "schema_url": schema_url,
+            "schema_type": schema_type
+        }
+
+        logger.info(f"🔍 Starting API schema analysis: {schema_url}")
+        result = hexstrike_client.safe_post("api/tools/api_schema_analyzer", data)
+
+        if result.get("success"):
+            analysis = result.get("schema_analysis_results", {})
+            endpoint_count = len(analysis.get("endpoints_found", []))
+            issue_count = len(analysis.get("security_issues", []))
+
+            logger.info(f"✅ Schema analysis completed: {endpoint_count} endpoints, {issue_count} issues")
+
+            if issue_count > 0:
+                logger.warning(f"⚠️  Found {issue_count} security issues in schema!")
+                for issue in analysis.get("security_issues", [])[:3]:  # Show first 3
+                    severity = issue.get("severity", "UNKNOWN")
+                    issue_type = issue.get("issue", "unknown")
+                    logger.warning(f"   ├─ [{severity}] {issue_type}")
+
+            if endpoint_count > 0:
+                logger.info(f"📊 Discovered endpoints:")
+                for endpoint in analysis.get("endpoints_found", [])[:5]:  # Show first 5
+                    method = endpoint.get("method", "GET")
+                    path = endpoint.get("path", "/")
+                    logger.info(f"   ├─ {method} {path}")
+        else:
+            logger.error("❌ Schema analysis failed")
+
+        return result
+
+    @mcp.tool()
+    def comprehensive_api_audit(base_url: str, schema_url: str = "", jwt_token: str = "", graphql_endpoint: str = "") -> Dict[str, Any]:
+        """
+        Comprehensive API security audit combining multiple testing techniques.
+
+        Args:
+            base_url: Base URL of the API
+            schema_url: Optional API schema URL
+            jwt_token: Optional JWT token for analysis
+            graphql_endpoint: Optional GraphQL endpoint
+
+        Returns:
+            Comprehensive audit results with all API security tests
+        """
+        audit_results = {
+            "base_url": base_url,
+            "audit_timestamp": time.time(),
+            "tests_performed": [],
+            "total_vulnerabilities": 0,
+            "summary": {},
+            "recommendations": []
+        }
+
+        logger.info(f"🚀 Starting comprehensive API security audit: {base_url}")
+
+        # 1. API Endpoint Fuzzing
+        logger.info("🔍 Phase 1: API endpoint discovery and fuzzing")
+        fuzz_result = api_fuzzer(base_url)
+        if fuzz_result.get("success"):
+            audit_results["tests_performed"].append("api_fuzzing")
+            audit_results["api_fuzzing"] = fuzz_result
+
+        # 2. Schema Analysis (if provided)
+        if schema_url:
+            logger.info("🔍 Phase 2: API schema analysis")
+            schema_result = api_schema_analyzer(schema_url)
+            if schema_result.get("success"):
+                audit_results["tests_performed"].append("schema_analysis")
+                audit_results["schema_analysis"] = schema_result
+
+                schema_data = schema_result.get("schema_analysis_results", {})
+                audit_results["total_vulnerabilities"] += len(schema_data.get("security_issues", []))
+
+        # 3. JWT Analysis (if provided)
+        if jwt_token:
+            logger.info("🔍 Phase 3: JWT token analysis")
+            jwt_result = jwt_analyzer(jwt_token, base_url)
+            if jwt_result.get("success"):
+                audit_results["tests_performed"].append("jwt_analysis")
+                audit_results["jwt_analysis"] = jwt_result
+
+                jwt_data = jwt_result.get("jwt_analysis_results", {})
+                audit_results["total_vulnerabilities"] += len(jwt_data.get("vulnerabilities", []))
+
+        # 4. GraphQL Testing (if provided)
+        if graphql_endpoint:
+            logger.info("🔍 Phase 4: GraphQL security scanning")
+            graphql_result = graphql_scanner(graphql_endpoint)
+            if graphql_result.get("success"):
+                audit_results["tests_performed"].append("graphql_scanning")
+                audit_results["graphql_scanning"] = graphql_result
+
+                graphql_data = graphql_result.get("graphql_scan_results", {})
+                audit_results["total_vulnerabilities"] += len(graphql_data.get("vulnerabilities", []))
+
+        # Generate comprehensive recommendations
+        audit_results["recommendations"] = [
+            "Implement proper authentication and authorization",
+            "Use HTTPS for all API communications",
+            "Validate and sanitize all input parameters",
+            "Implement rate limiting and request throttling",
+            "Add comprehensive logging and monitoring",
+            "Regular security testing and code reviews",
+            "Keep API documentation updated and secure",
+            "Implement proper error handling"
+        ]
+
+        # Summary
+        audit_results["summary"] = {
+            "tests_performed": len(audit_results["tests_performed"]),
+            "total_vulnerabilities": audit_results["total_vulnerabilities"],
+            "audit_coverage": "comprehensive" if len(audit_results["tests_performed"]) >= 3 else "partial"
+        }
+
+        logger.info(f"✅ Comprehensive API audit completed:")
+        logger.info(f"   ├─ Tests performed: {audit_results['summary']['tests_performed']}")
+        logger.info(f"   ├─ Total vulnerabilities: {audit_results['summary']['total_vulnerabilities']}")
+        logger.info(f"   └─ Coverage: {audit_results['summary']['audit_coverage']}")
+
+        return {
+            "success": True,
+            "comprehensive_audit": audit_results
+        }
